Make vcs.FromDir work for git submodules
This is a hacky fix that special cases looking for a .git file.
vcs.FromDir expected .git to be a directory.
However, for git submodules, their .git is a regular file that contains the
path to the git submodule's repo dir.

Ideally, we'd delegate to the vcs command and ask it if a given dir is a repo root.
This fix works for now, but is kinda hacky and not unit tested.